clean example
