+++ example/lib/spider.dart
@@ -1,0 +1,32 @@
+import 'package:dio/dio.dart';
+import 'package:html/parser.dart';
+import 'package:scrapy/scrapy.dart';
+
+import 'model.dart';
+
+class BlogSpider extends Spider<Quote, Quotes> {
+  Stream<String> Parse(Response response) async* {
+    var document = parse(response.data.toString());
+    var nodes = document.querySelectorAll("div.quote> span.text");
+
+    for (var node in nodes) {
+      yield node.innerHtml;
+    }
+  }
+
+  @override
+  Stream<String> Transform(Stream<String> stream) async* {
+    await for (String parsed in stream) {
+      var transformed = parsed;
+      yield transformed.substring(1, parsed.length - 1);
+    }
+  }
+
+  @override
+  Stream<Quote> Save(Stream<String> stream) async* {
+    await for (String transformed in stream) {
+      Quote quote = Quote(quote: transformed);
+      yield quote;
+    }
+  }
+}
